+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.11.0
 
--   `getCurrentRate` - returns an integer.
+-   `getCurrentRate` - always returns an integer.
 -   `pausable` - changed `proceed` to `maybeBlock`.
 
 # 0.10.0